+++ lib/src/algorithms/md4.dart
@@ -82,7 +82,7 @@ class MD4Hash extends BlockHashSink {
 
     for (int i = 16; i < 32; i++) {
       e = (b & c) | (b & d) | (c & d);
-      f = ((i >> 2) & 3) + ((i & 3) << 2);
+      f = ((i >>> 2) & 3) + ((i & 3) << 2);
       t = d;
       d = c;
       c = b;
@@ -94,7 +94,7 @@ class MD4Hash extends BlockHashSink {
 
     for (int i = 32; i < 48; i++) {
       e = b ^ c ^ d;
-      f = ((i & 1) << 3) | ((i & 2) << 1) | ((i >> 1) & 2) | ((i >> 3) & 1);
+      f = ((i & 1) << 3) | ((i & 2) << 1) | ((i >>> 1) & 2) | ((i >>> 3) & 1);
       t = d;
       d = c;
       c = b;